Reposition DocPopups to keep them onscreen
DocPopups positioned with position_at_location may hang offscreen.  This patch
attempts to keep this from happening.  Popups hanging off the bottom of the
screen are placed above the text in question, unless that would push it off
the top of the screen.  Popups hanging off the right of the screen are pushed
left until entirely onscreen, or until the popup reaches the left side of the
screen.

To do this, we need to know the size of the popup window.  Therefore, we have
to wait until after allocation to do the positioning.  position_at_location()
saves the information the positioning code to __position_info.  The actual
positioning code lives in do_size_allocate().